Fix MBC5 and minor changes to MBC trait

+++ mbc/src/mbc1.rs
@@ -156,8 +156,8 @@ impl MBCTrait for MBC1 {
         self.ram.clone()
     }
 
-    fn rom_banks(&self) -> u8 {
-        self.rom_banks
+    fn rom_banks(&self) -> u16 {
+        self.rom_banks as u16
     }
 
     fn ram_banks(&self) -> u8 {
